Detection of Entities type alias and GameObjects collection for systems

+++ src/Texture.cpp
@@ -36,11 +36,11 @@ bool Texture::LoadImpl( ) {
   stbi_set_flip_vertically_on_load( true );
   
   unsigned char *data{ stbi_load_from_memory( ( const unsigned char* )this->Data( ).c_str( ), 
-                                              this->Data( ).size( ), 
+                                              static_cast< int >( this->Data( ).size( ) ),
                                               &m_Width, 
                                               &m_Height, 
-                                              &m_NumChannels, 
-                                              0 
+                                              &m_NumChannels,
+                                              0
                                             )
   };
